requirements.txt

torch>=2.3.0
torchvision>=0.18.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.10.0
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
setuptools>=65.0.0